+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add fun buttons to spice up the mob farm
 * Add ability to configure mod item chest loot tables
+* Disable swimming logic on squids spawned by Clone-o-Matic
 
 ## 26.1.2.0